+++ apps/admin/src/components/aside-navigation-menu.tsx
@@ -12,6 +12,7 @@ import {
 
 import LOGO from '~/assets/logo.svg'
 import { PATH } from '~/constants/path'
+import { CLUB_ROUTE } from '~/features/club/constants'
 import { useRefreshTokens } from '~/hooks/use-refresh-token'
 import { useTokenExpiration } from '~/hooks/use-token-expiration'
 import { authServices } from '~/utils/auth'
@@ -35,7 +36,7 @@ const items: MenuItem[] = [
     icon: <GraduationCap size={20} />,
     children: [
       { key: 'dept', label: <Link to={PATH.DEPT}>학부 소개</Link> },
-      { key: 'club', label: <Link to={PATH.CLUB}>동아리 소개</Link> },
+      { key: 'club', label: <Link to={CLUB_ROUTE}>동아리 소개</Link> },
       {
         key: 'contact',
         label: <Link to={PATH.DIRECTIONS}>찾아오시는 길</Link>,
