+++ apps/admin/src/features/club/hooks/index.ts
@@ -2,4 +2,4 @@ export * from './useClubs'
 export * from './useCreateClub'
 export * from './useUpdateClub'
 export * from './useDeleteClub'
-export * from './useClubImageUPload'
+export * from './useImageUpload'
