+++ apps/admin/src/features/club/hooks/useDeleteClub.ts
@@ -1,6 +1,6 @@
 import { useQueryClient } from '@tanstack/react-query'
 import { message } from 'antd'
-import { useClubServiceDeleteApiV1ClubsById } from '~/apis/admin/queries'
+import { useClubServiceDeleteApiV1ClubsById } from '~/features/club/services'
 
 const CLUB_LIST_QK = ['ClubServiceGetApiV1Clubs'] as const
