+++ apps/admin/src/features/club/services/queries.ts
@@ -1,0 +1,10 @@
+
+export {
+  useClubServicePostApiV1Clubs,
+  useClubServicePatchApiV1ClubsById,
+  useClubServiceDeleteApiV1ClubsById,
+  useFileServicePostApiV1FilesClub,
+} from '~/apis/admin/queries'
+
+export { useClubServiceGetApiV1Clubs } from '~/apis/community/queries'
+export type { ClubDetailResponse } from '~/apis/community/requests'
