--- apps/admin/tsconfig.tsbuildinfo
@@ -1,1 +1,0 @@
-{"root":["./src/main.tsx","./src/routetree.gen.ts","./src/vite-env.d.ts","./src/apis/admin/openapiconfig.ts","./src/apis/admin/queries/common.ts","./src/apis/admin/queries/ensurequerydata.ts","./src/apis/admin/queries/index.ts","./src/apis/admin/queries/prefetch.ts","./src/apis/admin/queries/queries.ts","./src/apis/admin/queries/suspense.ts","./src/apis/admin/requests/index.ts","./src/apis/admin/requests/schemas.gen.ts","./src/apis/admin/requests/services.gen.ts","./src/apis/admin/requests/types.gen.ts","./src/apis/admin/requests/core/apierror.ts","./src/apis/admin/requests/core/apirequestoptions.ts","./src/apis/admin/requests/core/apiresult.ts","./src/apis/admin/requests/core/cancelablepromise.ts","./src/apis/admin/requests/core/openapi.ts","./src/apis/admin/requests/core/request.ts","./src/apis/community/queries/common.ts","./src/apis/community/queries/ensurequerydata.ts","./src/apis/community/queries/index.ts","./src/apis/community/queries/prefetch.ts","./src/apis/community/queries/queries.ts","./src/apis/community/queries/suspense.ts","./src/apis/community/requests/index.ts","./src/apis/community/requests/schemas.gen.ts","./src/apis/community/requests/services.gen.ts","./src/apis/community/requests/types.gen.ts","./src/apis/community/requests/core/apierror.ts","./src/apis/community/requests/core/apirequestoptions.ts","./src/apis/community/requests/core/apiresult.ts","./src/apis/community/requests/core/cancelablepromise.ts","./src/apis/community/requests/core/openapi.ts","./src/apis/community/requests/core/request.ts","./src/components/aside-navigation-menu.tsx","./src/components/club/club-creator.tsx","./src/components/club/club-table-view.tsx","./src/components/club/club-table.tsx","./src/components/hero-images/hero-carousel.tsx","./src/components/hero-images/hero-image-creator.tsx","./src/components/hero-images/editable-image-card/editable-image-card.tsx","./src/components/hero-images/editable-image-card/hero-image-manager.tsx","./src/components/lab/lab-creator.tsx","./src/components/lab/lab-table-view.tsx","./src/components/lab/lab-table.tsx","./src/components/main/posts-list-card.tsx","./src/components/posts/board.tsx","./src/components/posts/edit-post-field.tsx","./src/components/posts/posts-list.tsx","./src/components/posts/search-bar.tsx","./src/components/posts/write-new-post-field.tsx","./src/components/professor/professor-creator.tsx","./src/components/professor/professor-table-view.tsx","./src/components/professor/professor-table.tsx","./src/components/user/search-bar.tsx","./src/components/user/user-list.tsx","./src/constants/api.ts","./src/constants/path.ts","./src/hooks/use-edit-table.ts","./src/hooks/use-modal.ts","./src/hooks/use-refresh-token.ts","./src/hooks/use-sign-in.ts","./src/hooks/use-token-expiration.ts","./src/hooks/user/use-delete-users.ts","./src/routes/__root.tsx","./src/routes/index.tsx","./src/routes/club/index.tsx","./src/routes/dept/index.tsx","./src/routes/dept/edit/index.tsx","./src/routes/directions/index.tsx","./src/routes/directions/edit/index.tsx","./src/routes/hero-images/index.tsx","./src/routes/hero-images/edit/index.tsx","./src/routes/lab/index.tsx","./src/routes/main/index.tsx","./src/routes/news/$postid.tsx","./src/routes/news/index.tsx","./src/routes/news/edit/$postid.tsx","./src/routes/news/new/index.tsx","./src/routes/notice/$postid.tsx","./src/routes/notice/index.tsx","./src/routes/notice/edit/$postid.tsx","./src/routes/notice/new/index.tsx","./src/routes/professor/index.tsx","./src/routes/user/index.tsx","./src/utils/api.ts","./src/utils/auth.ts","./src/utils/get-query-client.ts","./src/utils/http.ts","./src/utils/jwt.ts","./src/utils/utils.ts"],"version":"5.8.3"}
